+++ examples/rtic_led.rs
@@ -30,7 +30,7 @@ mod app {
 
         // Device-specific peripherals
         let device: bsp::Peripherals = cx.device;
-        let pins = bsp::t40::into_pins(device.iomuxc);
+        let pins = bsp::t40::from_pads(device.iomuxc);
         let mut led = bsp::configure_led(pins.p13);
         led.set();