+++ examples/rtic_uart_log.rs
@@ -73,7 +73,7 @@ mod app {
             &mut cx.device.dcdc,
         );
 
-        let pins = bsp::t40::into_pins(cx.device.iomuxc);
+        let pins = bsp::t40::from_pads(cx.device.iomuxc);
 
         // UART setup.
         let uarts = cx.device.uart.clock(